# docs/api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "1Cati Site Management API",
    "version": "2026.07.01",
    "summary": "API contract for the 1Cati property-management web application.",
    "description": "This OpenAPI contract describes the local and production-ready API surface for the 1Cati CRM/ERP dashboard. It is used for API testing, Jira/Xray traceability and future client/provider handover."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3104",
      "description": "Local development server"
    },
    {
      "url": "https://cati-blond.vercel.app",
      "description": "Hosted web application"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AccessProfileCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_profile_role",
        "description": "Local QA access-profile cookie. Production uses Supabase Auth instead."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "SourceResponse": {
        "type": "object",
        "properties": {
          "source": { "type": "string", "enum": ["supabase", "local-seed"] }
        },
        "additionalProperties": true
      },
...
      "ActionLogRequest": {
        "type": "object",
        "required": ["actionType"],
        "properties": {
          "actionType": { "type": "string", "maxLength": 120 },
          "entityTable": { "type": "string" },
          "entityId": { "type": "string" },
          "entityExternalId": { "type": "string" },
          "title": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "ActionDecisionRequest": {
        "type": "object",
        "required": ["id", "status"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "status": { "type": "string", "enum": ["approved", "rejected", "completed", "failed"] },
          "actionType": { "type": "string", "maxLength": 120, "default": "ticket.update.request" },
          "entityTable": { "type": "string", "default": "service_tickets" }
        },
        "additionalProperties": false
      },
...
      "AiChatRequest": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "minLength": 1, "maxLength": 2000 }
        }
      },
...
      "DocumentUploadResponse": {
        "type": "object",
        "properties": {
          "storageMode": { "type": "string" },
          "upload": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": true
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "The user is not authenticated.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "The role is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerError": {
        "description": "The requested data or action is currently unavailable.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "security": [{ "AccessProfileCookie": [] }],
  "paths": {
...
    "/api/site-management/document-uploads": {
      "get": {
        "tags": ["Operations"],
        "operationId": "getDocumentUploadPolicy",
        "summary": "Return document upload policy.",
        "description": "Returns upload constraints, provider mode and role-specific upload policy.",
        "responses": { "200": { "description": "Document upload policy returned." }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["Operations"],
        "operationId": "uploadDocument",
        "summary": "Upload a document in simulation or live storage mode.",
        "description": "Accepts multipart form data, stores document metadata and returns upload status. Live storage requires configured provider credentials.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "title": { "type": "string" },
                  "category": { "type": "string" },
                  "flatNumber": { "type": "string" },
                  "packetId": { "type": "string" },
                  "note": { "type": "string" },
                  "retentionClass": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Document uploaded.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentUploadResponse" } } } },
          "400": { "description": "Invalid file or form data.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/site-management/actions": {
      "post": {
        "tags": ["Operations"],
        "operationId": "logClientAction",
        "summary": "Log an audited client action.",
        "description": "Stores an audited action after validating role permissions against the target entity and action type.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActionLogRequest" } } } },
        "responses": {
          "201": { "description": "Action logged.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceResponse" } } } },
          "400": { "description": "Invalid action body." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "patch": {
        "tags": ["Operations"],
        "operationId": "decideClientActionRequest",
        "summary": "Approve or reject an audited client action request.",
        "description": "Updates a pending action request after validating that the current role has approval permission for the resolved workflow action.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActionDecisionRequest" } } } },
        "responses": {
          "200": { "description": "Action request decision stored.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceResponse" } } } },
          "400": { "description": "Invalid action decision body." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/ai/chat": {
      "post": {
        "tags": ["AI"],
        "operationId": "postAiChat",
        "summary": "Return a role-aware AI assistant response.",
        "description": "Accepts a user message, detects language, applies RBAC guardrails and returns an AI or deterministic fallback response.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiChatRequest" } } } },
        "responses": {
          "200": { "description": "AI response returned." },
          "400": { "description": "Invalid message body." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "description": "Message is too long." }
        }
      }
    },
...
  }
}